Moved authentication stuff to new source file Closes #333
